backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.24.0
requests==2.31.0
pytest==7.4.2
pytest-flask==1.2.0